remove hack for auth headers
